refactor: 공통 Delete 전용 BaseEntity 생성
